zameenclone/requirements.txt
django_extensions
django_filter
django-fsm
faker~=26.0.0
pillow

djangorestframework~=3.15.2
djangorestframework-simplejwt

Django~=5.0.6
django-filter~=24.2
django-silk

celery~=5.4.0
sendgrid~=6.11.0
python_decouple~=3.8
django_celery_results~=2.5.1
django_celery_beat~=2.6.0
factory_boy~=3.3.0
